+++ env_mplab/src/UDP_app.h
@@ -61,7 +61,7 @@ SUBSTITUTE GOODS, TECHNOLOGY, SERVICES, OR ANY CLAIMS BY THIRD PARTIES
 #include "system_definitions.h"
 #include "adc.h"
 #define NB_UDP_INFO 7
-extern volatile uint16_t UDP_received_sample[NB_UDP_INFO];
+extern volatile uint8_t UDP_received_sample[NB_UDP_INFO];
 
 // *****************************************************************************
 // *****************************************************************************
